pyproject.toml
[project]
name = "jira-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "jira>=3.10.5",
    "mcp>=1.25.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
]
